+++ libs/core/src/test/java/org/opensearch/semver/expr/TildeTests.java
@@ -14,15 +14,16 @@ import org.opensearch.test.OpenSearchTestCase;
 public class TildeTests extends OpenSearchTestCase {
 
     public void testPatchVersionVariability() {
-        Tilde tildeExpr = new Tilde(Version.fromString("1.2.3"));
+        Tilde tildeExpr = new Tilde();
+        Version rangeVersion = Version.fromString("1.2.3");
 
-        assertTrue(tildeExpr.evaluate(Version.fromString("1.2.3")));
-        assertTrue(tildeExpr.evaluate(Version.fromString("1.2.4")));
-        assertTrue(tildeExpr.evaluate(Version.fromString("1.2.9")));
+        assertTrue(tildeExpr.evaluate(rangeVersion, Version.fromString("1.2.3")));
+        assertTrue(tildeExpr.evaluate(rangeVersion, Version.fromString("1.2.4")));
+        assertTrue(tildeExpr.evaluate(rangeVersion, Version.fromString("1.2.9")));
 
-        assertFalse(tildeExpr.evaluate(Version.fromString("1.2.0")));
-        assertFalse(tildeExpr.evaluate(Version.fromString("1.2.2")));
-        assertFalse(tildeExpr.evaluate(Version.fromString("1.3.0")));
-        assertFalse(tildeExpr.evaluate(Version.fromString("2.0.0")));
+        assertFalse(tildeExpr.evaluate(rangeVersion, Version.fromString("1.2.0")));
+        assertFalse(tildeExpr.evaluate(rangeVersion, Version.fromString("1.2.2")));
+        assertFalse(tildeExpr.evaluate(rangeVersion, Version.fromString("1.3.0")));
+        assertFalse(tildeExpr.evaluate(rangeVersion, Version.fromString("2.0.0")));
     }
 }